cleaning up code
This is using the new hero cfg's, this is using two family of variables, range, and up.
Range'0-5' is the range to be displayed on ability usage, only on cheat enabled lobbies.
up'0-5''0-3' is the command that gets executed when the player levels up. Defaulting to an echo of the number it is.

An example of both are:
range3
up52

I also added a function that toggles all auto-casts, so that there are less hero dependent functions.
And 3 different ways to view the runes, cycle, move to, and move to and back (to hero).

I am planning on adding more functions, I just need to find what needs adding.

I kinda ignored the A, AA, AQ, I and IQ, as they are just boring constants of the using abilities or items. One glance should suffice to an understanding.