Re-write for new content.
I can't convey words well, and my spell checker wasn't working. There will be grammatical and spelling mistakes. If you could help me find them I would like that! If you want to rewrite this go ahead!

I didn't find a need to credit the original producer, as it is a full rewrite. However, if you do, then please by all means add a link, I would do it in the 'how it works', as I use a modified version of their loop, so that it works better. And after finding out what it was used for, it's rather cool!

I also split it up a bit more then it already is, so that it's easier to install and use.